@@ -12,11 +12,14 @@
     <url>http://topjava.herokuapp.com/</url>
 
     <properties>
-        <java.version>1.8</java.version>
+        <java.version>17</java.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
 
         <spring.version>5.3.15</spring.version>
+        <spring-data-jpa.version>2.6.2</spring-data-jpa.version>
+
+        <tomcat.version>9.0.59</tomcat.version>
 
         <!-- Logging -->
         <logback.version>1.2.10</logback.version>
